some locks and cleaning up

+++ inc/memlayout.h
@@ -180,7 +180,7 @@ struct Page {
   // Pages allocated at boot time using pmap.c's
   // boot_alloc do not have valid reference count fields.
 
-  uint16_t pp_ref;
+  uint32_t pp_ref;
 };
 
 #endif                          /* !__ASSEMBLER__ */
